Extract risk attribution contracts
